+++ src/views/DiariasView.jsx
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import { exportCSV, exportODS, exportPDF, ExportMenu } from "../exportHelpers.jsx";
+import { ExportMenu } from "../exportHelpers.jsx";
 import { clickable } from "../utils.js";
 import FilterBar from "../components/FilterBar.jsx";